refactor(core): make class and function reflection introspection more unified and consistent, use common introspection context class
