+++ kotlinx-schema-generator-core/src/jvmMain/kotlin/kotlinx/schema/generator/reflect/ReflectionFunctionIntrospector.kt
@@ -7,7 +7,6 @@ import kotlinx.schema.generator.core.ir.TypeGraph
 import kotlinx.schema.generator.core.ir.TypeId
 import kotlinx.schema.generator.core.ir.TypeRef
 import kotlin.reflect.KCallable
-import kotlin.reflect.KClass
 import kotlin.reflect.KParameter
 
 /**
@@ -30,7 +29,7 @@ public object ReflectionFunctionIntrospector : SchemaIntrospector<KCallable<*>, 
             "Extension functions are not supported"
         }
 
-        val context = ReflectionClassIntrospector.IntrospectionContext()
+        val context = ReflectionIntrospectionContext()
 
         // Extract function information
         val functionName = root.name
@@ -48,10 +47,10 @@ public object ReflectionFunctionIntrospector : SchemaIntrospector<KCallable<*>, 
             val paramType = param.type
             val hasDefault = param.isOptional
 
-            val typeRef = context.convertKTypeToTypeRef(paramType)
+            val typeRef = context.toRef(paramType)
 
             // Extract description from annotations
-            val description = extractDescription(param.annotations)
+            val description = context.extractDescription(param.annotations)
 
             properties +=
                 Property(
@@ -71,87 +70,12 @@ public object ReflectionFunctionIntrospector : SchemaIntrospector<KCallable<*>, 
                 name = functionName,
                 properties = properties,
                 required = requiredProperties,
-                description = extractDescription(root.annotations),
+                description = context.extractDescription(root.annotations),
             )
 
         // Add an object generated from a function to the nodes
         val nodes = context.nodes + (id to objectNode)
 
         return TypeGraph(root = TypeRef.Ref(id, nullable = false), nodes = nodes)
-    }
-
-    /**
-     * Maintains state during function introspection including discovered nodes,
-     * visited classes, and type reference cache.
-     */
-    private class IntrospectionContext : ReflectionIntrospectionContext() {
-        private val defaultValueExtractor = DefaultValueExtractor()
-
-        /**
-         * Converts a KCallable (function) to a TypeRef representing its parameters as an object.
-         */
-        fun convertFunctionToTypeRef(callable: KCallable<*>): TypeRef {
-            val functionName = callable.name
-            val id = TypeId(functionName)
-
-            // Create an ObjectNode representing the function parameters
-            val properties = mutableListOf<Property>()
-            val requiredProperties = mutableSetOf<String>()
-
-            callable.parameters.forEach { param ->
-                // Skip instance parameter for member functions
-                if (param.kind == KParameter.Kind.INSTANCE) return@forEach
-
-                val paramName = param.name ?: return@forEach
-                val paramType = param.type
-                val hasDefault = param.isOptional
-
-                val typeRef = convertKTypeToTypeRef(paramType)
-
-                // Extract description from annotations
-                val description = extractDescription(param.annotations)
-
-                properties +=
-                    Property(
-                        name = paramName,
-                        type = typeRef,
-                        description = description,
-                        hasDefaultValue = hasDefault,
-                    )
-
-                if (!hasDefault) {
-                    requiredProperties += paramName
-                }
-            }
-
-            val objectNode =
-                ObjectNode(
-                    name = functionName,
-                    properties = properties,
-                    required = requiredProperties,
-                    description = extractDescription(callable.annotations),
-                )
-
-            _nodes[id] = objectNode
-            return TypeRef.Ref(id, nullable = false)
-        }
-
-        override fun createObjectNode(
-            klass: KClass<*>,
-            parentPrefix: String?,
-        ): ObjectNode {
-            // Try to extract default values by creating an instance
-            val defaultValues = defaultValueExtractor.extractDefaultValues(klass)
-
-            // Extract properties from primary constructor using shared method
-            val (properties, requiredProperties) = extractConstructorProperties(klass, defaultValues)
-
-            return ObjectNode(
-                name = klass.simpleName ?: "UnknownClass",
-                properties = properties,
-                required = requiredProperties,
-                description = extractDescription(klass.annotations),
-            )
-        }
     }
 }
